Выводит поинты неотрывно от дней
