+++ src/components/trip-day.js
@@ -6,8 +6,8 @@ const createTripDayTemplate = (tripDayInfo) => {
   return (
     `<li class="trip-days__item  day">
         <div class="day__info">
-          <span class="day__counter">${correctDayFormat(tripDayInfo.date)}</span>
-          <time class="day__date" datetime="${correctDateFormat(tripDayInfo.date)}">${correctMonthAndYearFormat(tripDayInfo.date)}</time>
+          <span class="day__counter">${correctDayFormat(tripDayInfo)}</span>
+          <time class="day__date" datetime="${correctDateFormat(tripDayInfo)}">${correctMonthAndYearFormat(tripDayInfo)}</time>
         </div>
         <ul class="trip-events__list"></ul>
       </li>`
