+++ src/controllers/trip-days.js
@@ -8,88 +8,79 @@ import NoPointsComponent from '../components/no-points.js';
 
 export const EmptyPoint = {};
 
+const getDays = (points) => {
+  const daysList = [];
+  const days = new Set(points.map((point) => point.departure.getTime()));
+
+  for (const day of days) {
+    daysList.push(new Date(day));
+  }
+
+  return daysList;
+};
+
 export default class TripController {
   constructor(container, pointsModel) {
     this._container = container;
     this._pointsModel = pointsModel;
 
+    this._sortComponent = new SortComponent();
     this._tripDaysComponent = new TripDaysComponent();
+
+    this._points = null;
     this._tripDayComponent = null;
 
-    this._sortComponent = new SortComponent();
     this._showedPointsControllers = [];
 
     this._onDataChange = this._onDataChange.bind(this);
     this._onViewChange = this._onViewChange.bind(this);
-    this._onFilterChange = this._onFilterChange.bind(this);
-
-    this._pointsModel.setFilterChangeHandler(this._onFilterChange);
+    // this._onFilterChange = this._onFilterChange.bind(this);
+    //
+    // this._pointsModel.setFilterChangeHandler(this._onFilterChange);
   }
 
-  render(days) {
-    this._days = days;
-    const points = this._pointsModel.getPointsAll();
+  render() {
+    this._points = this._pointsModel.getPointsAll();
+    console.log(this._points);
+    // Пороверка точек маршрута на наличие;
+    const isAllPointsMissing = this._points.every((point) => point.length === 0);
 
-    // Пороверка точек маршрута на наличие
-    const isAllPointsMissing = points.every((point) => point.length === 0);
-    const isAllDaysMissing = days.every((day) => day.length === 0);
-
-    if (isAllPointsMissing || isAllDaysMissing) {
+    if (isAllPointsMissing) {
       render(this._container, new NoPointsComponent());
       return;
     }
 
-    // Отрисовка меню сортировки
+    // Отрисовка меню сортировки;
     render(this._container, this._sortComponent);
 
-    // Отрисовка "контейнера" для вывода всех дней путешествия
+    // Отрисовка "контейнера" для вывода всех дней путешествия;
     render(this._container, this._tripDaysComponent);
 
-    // Отрисовка дней путешествия
-    this._renderTripDays();
+    // Отрисовка дней путешествия и точек маршрута;
+    this._renderPoints(this._points);
 
-    // Отрисовка точек маршрута в днях путешествия
-    this._renderDaysTripPoints();
 
-    // Сортировка точек маршрута в зависимости от выбранного параметра
-    const getSortedTrips = (sortType) => {
 
-      const renderSortPoints = (sortPointsList) => {
-        this._tripDayComponent = new TripsContainerComponent();
-        render(this._tripDaysComponent.getElement(), this._tripDayComponent);
-        this._renderPoints(this._tripDayComponent.getElement().querySelector(`.trip-events__list`), sortPointsList);
-      };
+  }
 
-      switch (sortType) {
-        case SortTypes.SORT_EVENT:
-          this._tripDaysComponent.getElement().innerHTML = ``;
+  // Отрисовка точек маршрута в днях путешествия
+  _renderPoints(points) {
+    const days = getDays(points);
+    // const pointsContainerElement = this._tripDaysComponent.getElement();
 
-          this._renderTripDays();
-          this._renderDaysTripPoints();
-          break;
+    for (const day of days) {
+      this._tripDayComponent = new TripDayComponent(day);
+      render(this._tripDaysComponent.getElement(), this._tripDayComponent);
 
-        case SortTypes.SORT_TIME:
-          const tripPointsListTime = points.slice();
-          tripPointsListTime.sort((a, b) => a.arrival - a.departure < b.arrival - b.departure ? 1 : -1);
-          this._tripDaysComponent.getElement().innerHTML = ``;
-
-          renderSortPoints(tripPointsListTime);
-          break;
-
-        case SortTypes.SORT_PRICE:
-          const tripPointsListPrice = points.slice();
-          tripPointsListPrice.sort((a, b) => a.price < b.price ? 1 : -1);
-          this._tripDaysComponent.getElement().innerHTML = ``;
-
-          renderSortPoints(tripPointsListPrice);
-          break;
+      for (const point of points) {
+        if (point.departure.getTime() === day.getTime()) {
+          const tripEventsListElement = this._tripDayComponent.getElement().querySelector(`.trip-events__list`);
+          const pointController = new PointController(tripEventsListElement, this._onDataChange, this._onViewChange);
+          pointController.render(point, Mode.DEFAULT);
+          this._showedPointsControllers = this._showedPointsControllers.concat(pointController);
+        }
       }
-    };
-
-    // Обрботка клика по кнопкам меню сортировки
-    this._sortComponent.setSortTypeChangeHandler(() => {
-      getSortedTrips(this._sortComponent.getSortType());
-    });
+    }
   }
 
   _onDataChange(pointController, oldData, newData) {
@@ -116,51 +107,4 @@ export default class TripController {
   _onViewChange() {
     this._showedPointsControllers.forEach((it) => it.setDefaultView());
   }
-
-  // Отрисовка дней путешествия
-  _renderTripDays() {
-    for (let i = 0; i < this._days.length; i++) {
-      this._tripDayComponent = new TripDayComponent(this._days[i]);
-      render(this._tripDaysComponent.getElement(), this._tripDayComponent);
-    }
-  }
-
-  // Отрисовка точек маршрута в днях путешествия
-  _renderDaysTripPoints() {
-    const tripEventsListElements = this._tripDaysComponent.getElement().querySelectorAll(`.trip-events__list`);
-
-    for (let i = 0; i < this._days.length; i++) {
-      const currentTripDay = tripEventsListElements[i];
-      const wayPoint = this._days[i].wayPoints;
-
-      this._renderPoints(currentTripDay, wayPoint);
-    }
-  }
-
-  _renderPoints(container, points) {
-    // const pointsContainerElement = this._tripDaysComponent.getElement().querySelector(`.trip-events__list`);
-    for (const point of points) {
-      const pointController = new PointController(container, this._onDataChange, this._onViewChange);
-      pointController.render(point, Mode.DEFAULT);
-      this._showedPointsControllers = this._showedPointsControllers.concat(pointController);
-    }
-  }
-
-  _removePoints() {
-    this._tripDaysComponent.getElement().innerHTML = ``;
-    this._showedPointsControllers.forEach((pointController) => pointController.destroy());
-    this._showedPointsControllers = [];
-  }
-
-  _updatePoints() {
-    this._removePoints();
-    this._renderPoints(this._tripDaysComponent.getElement(), this._pointsModel.getPoints());
-
-    console.log(this._pointsModel.getPoints());
-  }
-
-  _onFilterChange() {
-    this._updatePoints();
-  }
-
 }
